+++ test/unit/resolveFilename.js
@@ -27,54 +27,18 @@ describe('resolveFilename', function(){
     expect(result.relative).to.match(/.*opal-browser.rb/)
   })
 
-  it('allows additional MRI requires', function() {
-    process.env.OPAL_MRI_REQUIRES = 'opal-browser'
+  it('resolves corelib/runtime as a JS file in Bundler mode', function() {
+    const result = bundlerResolve('corelib/runtime')
 
-    const result = resolveFilename('opal-browser')
-
-    expect(result.absolute).to.match(/gems\/opal-browser-.*\/opal\/opal-browser\.rb/)
-    expect(result.relative).to.match(/.*opal-browser.rb/)
+    expect(result.absolute).to.match(/gems\/opal.*corelib\/runtime.js/)
+    expect(result.relative).to.match(/.*corelib\/runtime.js/)
   })
 
-  it('allows additional MRI requires with bundler', function() {
-    process.env.OPAL_MRI_REQUIRES = 'opal-builder'
-
-    const result = bundlerResolve('builder')
-
-    expect(result.absolute).to.match(/gems\/opal-builder-.*\/opal\/builder\.rb/)
-    expect(result.relative).to.match(/.*builder.rb/)
-  })
-
-  it('passes bundled copy of opal through in non bundler mode', function() {
+  it('resolves the Opal compiler in non bundler mode', function() {
     const result = resolveFilename('opal')
 
     expect(result.absolute).to.eq(path.resolve(__dirname, '../../vendor/opal-compiler.js'))
     expect(result.relative).to.eq('opal')
-  })
-
-  it('Rails hook is listening', function() {
-    process.env.RAILS_ENV = 'unexpected_env'
-
-    expect(function() { bundlerResolve('opal-browser')}).to.throw(/Command failed: rails runner/)
-  })
-
-  it('uses Rails load paths if Rails is running', function() {
-    process.env.RAILS_ENV = 'foobar'
-
-    const result = bundlerResolve('opal-browser')
-
-    expect(result.absolute).to.match(/gems\/opal-browser-.*\/opal\/opal-browser\.rb/)
-    expect(result.relative).to.match(/.*opal-browser.rb/)
-  })
-
-  it('allows additional MRI requires with Rails', function() {
-    process.env.RAILS_ENV = 'foobar'
-    process.env.OPAL_MRI_REQUIRES = 'opal-builder'
-
-    const result = resolveFilename('builder')
-
-    expect(result.absolute).to.match(/gems\/opal-builder-.*\/opal\/builder\.rb/)
-    expect(result.relative).to.match(/.*builder.rb/)
   })
 
   it('resolves a test fixture', function() {